Added bitwise tests
